Move DataRow to generic location for reuse
